decktracker: change behavior of tooltip
It now shows only the user stays a bit on the card and moves the mouse.
Not ideal, but I didn't have another solution to cancel the tooltip
if the user mouves direclty from the game card to outside the game
window, as no "mouseleave" event is fired in this case.