DEV: add the ability for events to trigger a reset of history
Today, the history is the view + the data linked to the view.
While this seemed a good idea at first (super easy to move
through the history), it has a big issue when new data arrives,
since going back in history + breaking the history by doing something
else wipes the data for the duration of the session.
Ideally, the history should be concerned only about location, and
then have a separate process compute the data in this location based
on the (non-historized) reference data