DEV: do a gradual load for cards in cards display list
Try and reduce the perf hit when displaying the full page.

I made some improvements, but it's still not as smooth
as what we would like.
I did tests using Chrome Performance tools, setting a Slow 3G
internet speed and a 6x slowdown of the CPU to really see how
the perf is impacted.

The performance hit comes when loading and displaying the image.
So what I'm doing is to only load a few at a time, then pause a
bit, so that the UI is not completely stuck.

However, if we want it to be super smooth (on low end computers),
it means more or less loading the images one by one with a big
pause in-between, which makes the collection page more or less
useless (it has value only once all the images are loaded - if
the user knows what card they want, they will use the search
instead).

So I don't have a perfect solution for now. Please test the new
implementation and let me know if you feel a difference or not
at least