DEV: remove some prefs out of the navigation state
Prefs in the navigation state don't work, as if you
go back in the history you would then get the old
pref.
The state of the app right now is not really good.
Some prefs are handled in the main state, and some
prefs are simply forwarded via the prefs event bus.
I'm not sure yet which approach is the best one, so
for now the code will stay in this awkward state.