BUG (tracker): fix deck sometimes not updating for friendly matches
This changes completely the paradigm of how to get an active
deck from the memory, and the memory updates seems to always
have to reset after reaching the deck selection screen for the
first time. It seems to be working, and the reset only happens
once, so for now I'll test it like this.