BUG (bgs): try and fix some issues with long-running games
One of the issues is that sometimes, the PTL logs get truncated,
and "options" are sent which force the app to go back to the root
nodes.
This usually works, but because in BG some big parts of the logs
get removed, it would occasionally break the parser.
The fix removes the "go back to root" in PTL parsing. This might in turn
introduces other issues down the line however, so we'll have to
keep watch of them.