BUG (tracker): always use the "base card id" when manipulating the deck
So that rank 2 and rank 1 cards are treated the same way.
This might cause an issue with upgraded spellstones that have
been sent back to the deck though, as we will treat them the
same way as the unupgraded one. However, in that case we have
the entity ID, so maybe we're fine