Some cosmetic improvements
Increased the size of the "Copy deck code" button. Now not only the icon is clickable, but also the label itself.
Removed duplicate class names in tooltip for neutral heroes in Duels. Also changed the tooltip, so now it's "Hero (Class)"
Added a null check for the player's MMR in Duels. (I actually encountered null instead of my MMR)
Global-stat blocks is set to a fixed height. For the case when in non-English languages the string (stat.label) takes two lines which causes the blocks to have different heights. (The numerical value of the height may need to be changed.)
Label "Max level!" in Mercs->Progress center aligned (for non-English languages, where it takes two lines)
Improved BG hero power text sanitizing (for Lich Bazhial)
Standard and Wild packs have been moved to the main sets. So for now it's true, that special packs is unbuyable ones.
I hope I fixed reset-button-tooltip