BUG (tracker): fix stealing entity from deck not working
The main issue was that only the entityId was used to find
the stolen entity. I can't be sure exactly why (thank you
past me for not adding comments), but my guess is that
using the cardId could lead to us picking the wrong card
to steal, because it would look through all the zones.
This fix now has the app look into the correct zone right
away, and then use the cardId to be able to fallback if no
card with the matching entityId is found.