DEV: extract game info memory reading events provider to own service
The idea is to gradually move everything towards a reactive model,
so that we can easily populate the information as it arrives. This
should simplify the end-game uploader and reduce the number of errors
caused by missing data (because we won't be in a rush to get it).
Maybe this will also help duels run detection